feat: creates /src folder, routes.py file and blueprints
