+++ .gitignore
@@ -1,4 +1,3 @@
-venv
-.env
+**env
 **/__pycache__
 .vscode
